refactor: force .toPacket() to be called on player info before passing it into the message queue in session; this ensures the .toPacket() is effectively called from the same engine threads, and the correct reference is passed onto netty threads to use
